Struct samd_dma::DMAController

pub struct DMAController<T: 'static + DmaStorage> { /* fields omitted */ }

DMA system controller.

Used to distribute channels, as well as control higher level operations of the DMA system.

Methods

impl<T: 'static + DmaStorage> DMAController<T>

pub fn init(dmac: DMAC, storage: &'static mut T) -> DMAController<T>

Initialise the DMA Controller with the specified storage.

pub fn disable(&mut self)

Disable all channels and the CRC module. This will abort any ongoing DMA transactions. The DMA system will not be fully disabled until any ongoing burst transfer is completed.

pub fn enable(&mut self)

Enable the DMA system.

pub fn is_enabled(&self) -> bool

Returns true if the DMA system is enabled.

pub fn take_channel<U: Unsigned>(&mut self) -> Option<Channel> where
U: IsLess<T::Index, Output = True>,

Take a DMA channel by its number. If the channel has already been taken or if the ID is not available, None is returned. If you don't call return_channel on this channel, you can never get it back.

Safety

Taking the same channel from overlapping interrupt contexts could lead to double channels all the way. This will alias the exclusive references of the base and write-back descriptors corresponding to this channel. The same problem applies when a call to return_channel overlaps with a call to this function for the same channel.

pub fn return_channel(&mut self, channel: Channel)

Return a channel to the controller. This will disable the channel and reset the channel and first transfer descriptor.

Safety

The same interrupt safety issues with take_channel apply here as well.

pub fn enable_priority_level(&mut self, level: Priority)

Allow channels with the corresponding priority level to be part of arbitration.

pub fn diable_priority_level(&mut self, level: Priority)

Deny channels with the corresponding priority level to be a part of arbitration.

pub fn priority_level_enabled(&self, level: Priority) -> bool

Return true if the priority level is enabled.

pub fn set_priority_level_scheduling(&mut self, level: Priority, enable: bool)

Enable or disable round-robin scheduling method for channels of the given priority level. Disabling round-robin scheduling will enable static scheduling.